+++ Source/NSPropertyList.m
@@ -83,7 +83,7 @@ static Class	NSMutableStringClass;
 static Class	GSStringClass;
 static Class	GSMutableStringClass;
 
-extern BOOL GSScanDouble(unichar*, unsigned, double*);
+extern BOOL GSScanDouble(unichar*, NSUInteger, double*);
 
 @class	GSMutableDictionary;
 @interface GSMutableDictionary : NSObject	// Help the compiler
@@ -791,12 +791,12 @@ static inline id parseQuotedString(pldata* pld)
     }
   else
     {
-      unsigned	length;
-      unichar	*chars;
-      unichar	*temp = NULL;
-      unsigned	int temp_length = 0;
-      unsigned	j;
-      unsigned	k;
+      NSUInteger  length;
+      unichar     *chars;
+      unichar     *temp = NULL;
+      NSUInteger  temp_length = 0;
+      NSUInteger  j;
+      NSUInteger  k;
 
       if (!GSToUnicode(&temp, &temp_length, &pld->ptr[start],
 		       pld->pos - start, NSUTF8StringEncoding,
